clean steakhouse docs
