+++ docs/steakhouse/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "site",
+  "name": "steakhouse-docs",
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "vocs dev",
+    "dev": "vocs dev --port 5173",
     "build": "vocs build",
     "preview": "vocs preview"
   },
