+++ docs/steakhouse/pages/index.mdx
@@ -1,0 +1,15 @@
+---
+layout: landing
+---
+
+import { HomePage } from 'vocs/components'
+
+<HomePage.Root>
+  <HomePage.Logo />
+  <HomePage.Tagline>Type-safe SDK for [Steakhouse](https://steakhouse.financial)-curated DeFi vaults.</HomePage.Tagline>
+  <HomePage.InstallPackage name="@whisk/steakhouse" type="add" />
+  <HomePage.Buttons>
+    <HomePage.Button href="/getting-started" variant="accent">Get started</HomePage.Button>
+    <HomePage.Button href="https://github.com/papercliplabs/whisk-sdk">GitHub</HomePage.Button>
+  </HomePage.Buttons>
+</HomePage.Root>
