+++ docs/steakhouse/tsconfig.json
