+++ docs/whisk/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "whisk-docs",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vocs dev --port 5174",
+    "build": "vocs build",
+    "preview": "vocs preview"
+  },
+  "dependencies": {
+    "react": "catalog:",
+    "react-dom": "catalog:",
+    "vocs": "^1.4.1"
+  },
+  "devDependencies": {
+    "@types/react": "catalog:"
+  }
+}
