+++ docs/whisk/public/graphql-example.png
