+++ docs/whisk/public/how-it-works.png
