+++ docs/whisk/public/icon-dark.png
