+++ docs/whisk/public/icon-light.png
