+++ packages/client/src/WhiskClient.test.ts
@@ -1,16 +1,16 @@
 import { describe, expect, it } from "vitest"
-import { SteakhouseClient } from "./client.js"
+import { WhiskClient } from "./WhiskClient.js"
 
-describe("SteakhouseClient", () => {
+describe("WhiskClient", () => {
   it("creates client with config", () => {
-    const client = new SteakhouseClient({
+    const client = new WhiskClient({
       apiKey: "test-key",
     })
     expect(client).toBeDefined()
   })
 
   it("creates client with custom url", () => {
-    const client = new SteakhouseClient({
+    const client = new WhiskClient({
       apiKey: "test-key",
       url: "https://custom.api.example.com/graphql",
     })
@@ -18,7 +18,7 @@ describe("SteakhouseClient", () => {
   })
 
   it("creates client with debug mode", () => {
-    const client = new SteakhouseClient({
+    const client = new WhiskClient({
       apiKey: "test-key",
       debug: true,
     })
