+++ packages/steakhouse/src/index.ts
@@ -1,3 +1,2 @@
-export type { WhiskClientConfig as SteakhouseClientConfig } from "@whisk/client"
-export { createSteakhouseClient, SteakhouseClient } from "./client.js"
+export * from "./client.js"
 export * from "./queries/index.js"
