+++ packages/steakhouse/src/queries/index.ts
@@ -1,24 +1,10 @@
 // Fragments
 export { type VaultDetail, vaultDetailFragment } from "./fragments/vaultDetail.js"
-
-// Vault queries
-export {
-  buildSteakhouseMetadata,
-  type DetailedVault,
-  type GetDetailedVaultResult,
-  type GetDetailedVaultVariables,
-  getDetailedVault,
-  type SteakhouseMetadata,
-  vaultQuery as vaultDetailQuery,
-  vaultWithHistoricalQuery,
-} from "./getVault.js"
-export {
-  type GetVaultsResult,
-  type GetVaultsVariables,
-  getVaults,
-  type VaultWithMetadata,
-  vaultsQuery,
-} from "./getVaults.js"
+// Queries
+export * from "./getDetailedVault.js"
+export * from "./getVaults.js"
+// Shared
+export { buildSteakhouseMetadata, type SteakhouseMetadata } from "./steakhouseMetadata.js"
 
 // Types
 export type { SteakhouseQueryFn } from "./types.js"
