+++ packages/steakhouse/src/react/index.ts
@@ -1,7 +1,2 @@
 export * from "./hooks/index.js"
-export {
-  type SteakhouseContextValue,
-  SteakhouseProvider,
-  type SteakhouseProviderProps,
-  useSteakhouse,
-} from "./provider.js"
+export * from "./provider.js"
