+++ pnpm-workspace.yaml
@@ -1,7 +1,7 @@
 packages:
   - packages/*
   - examples/*
-  - site
+  - docs/*
 
 catalog:
   '@tanstack/react-query': 5.90.10
